+++ Scripts/lint.sh
@@ -33,6 +33,10 @@ check_package_info_plist() {
   "${ROOT_DIR}/Scripts/test_package_info_plist.sh"
 }
 
+check_package_runtime_integrity() {
+  "${ROOT_DIR}/Scripts/test_package_runtime_integrity.sh"
+}
+
 check_release_dsym_paths() {
   "${ROOT_DIR}/Scripts/test_release_dsym_paths.sh"
 }
@@ -89,6 +93,7 @@ run_portable_checks() {
   check_package_strip
   check_package_signing
   check_package_info_plist
+  check_package_runtime_integrity
   check_release_dsym_paths
   check_sparkle_signing_paths
   check_swift_test_sharding